+++ Engine/src/pragmas.c
@@ -4,56 +4,56 @@
 #include "platform.h"
 #include "pragmas.h"
 
-unsigned long getkensmessagecrc(long param) {
+uint32_t getkensmessagecrc(int32_t param) {
     return(0x56c764d4);
 }
 
 
-void clearbuf(void *d, long c, long a)
+void clearbuf(void *d, int32_t c, int32_t a)
 {
-	long *p = (long*)d;
+	int32_t *p = (long*)d;
 	while ((c--) > 0) *(p++) = a;
 }
 
-void clearbufbyte(void *D, long c, long a)
+void clearbufbyte(void *D, int32_t c, int32_t a)
 { // Cringe City
 	char *p = (char*)D;
-	long m[4] = { 0xffl,0xff00l,0xff0000l,0xff000000l };
-	long n[4] = { 0,8,16,24 };
-	long z=0;
+	int32_t m[4] = { 0xffl,0xff00l,0xff0000l,0xff000000l };
+	int32_t n[4] = { 0,8,16,24 };
+	int32_t z=0;
 	while ((c--) > 0) {
 		*(p++) = (char)((a & m[z])>>n[z]);
 		z=(z+1)&3;
 	}
 }
 
-void copybuf(void *s, void *d, long c)
+void copybuf(void *s, void *d, int32_t c)
 {
-	long *p = (long*)s, *q = (long*)d;
+	int32_t *p = (long*)s, *q = (long*)d;
 	while ((c--) > 0) *(q++) = *(p++);
 }
 
-void copybufbyte(void *S, void *D, long c)
+void copybufbyte(void *S, void *D, int32_t c)
 {
 	char *p = (char*)S, *q = (char*)D;
 	while((c--) > 0) *(q++) = *(p++);
 }
 
-void copybufreverse(void *S, void *D, long c)
+void copybufreverse(void *S, void *D, int32_t c)
 {
 	char *p = (char*)S, *q = (char*)D;
 	while((c--) > 0) *(q++) = *(p--);
 }
 
-void qinterpolatedown16(long* bufptr, long num, long val, long add)
+void qinterpolatedown16(long* bufptr, int32_t num, int32_t val, int32_t add)
 { // gee, I wonder who could have provided this...
-    long i, *lptr = bufptr;
+    int32_t i, *lptr = bufptr;
     for(i=0;i<num;i++) { lptr[i] = (val>>16); val += add; }
 }
 
-void qinterpolatedown16short(long* bufptr, long num, long val, long add)
+void qinterpolatedown16short(long* bufptr, int32_t num, int32_t val, int32_t add)
 { // ...maybe the same person who provided this too?
-    long i; short *sptr = (short *)bufptr;
+    int32_t i; short *sptr = (short *)bufptr;
     for(i=0;i<num;i++) { sptr[i] = (short)(val>>16); val += add; }
 }
